// dashboard/src/utils/hasura-api/generated/schemas/trackFunctionBulkOperation.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { TrackFunctionBulkOperationType } from './trackFunctionBulkOperationType';
import type { TrackFunctionStep } from './trackFunctionStep';

export interface TrackFunctionBulkOperation {
  type: TrackFunctionBulkOperationType;
  source?: string;
  /** Resource version of the operation */
  resource_version: number;
  args: TrackFunctionStep[];
}