// dashboard/src/utils/hasura-api/generated/schemas/trackFunctionStep.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { TrackFunctionArgs } from './trackFunctionArgs';
import type { TrackFunctionStepType } from './trackFunctionStepType';

export interface TrackFunctionStep {
  type: TrackFunctionStepType;
  args: TrackFunctionArgs;
}