// dashboard/src/utils/hasura-api/generated/schemas/untrackFunctionStep.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { UntrackFunctionArgs } from './untrackFunctionArgs';
import type { UntrackFunctionStepType } from './untrackFunctionStepType';

export interface UntrackFunctionStep {
  type: UntrackFunctionStepType;
  args: UntrackFunctionArgs;
}